+++ gala/potential/potential/tests/test_util.py
@@ -1,7 +1,9 @@
+import pytest
+
 # This project
-from ....units import solarsystem
 from ..util import from_equation
 from .helpers import PotentialTestBase
+
 
 class EquationBase(PotentialTestBase):
     def test_plot(self):
@@ -16,16 +18,22 @@ class EquationBase(PotentialTestBase):
         # Skip for now because these can't be written to YAML
         pass
 
+
 class TestHarmonicOscillatorFromEquation(EquationBase):
     Potential = from_equation("1/2*k*x**2", vars="x", pars="k",
                               name='HarmonicOscillator',
                               hessian=True)
     potential = Potential(k=1.)
-    w0 = [1.,0.]
+    w0 = [1., 0.]
 
     def test_derp(self):
         import numpy as np
         self.potential.gradient(np.random.random(size=(1,13)))
+
+    @pytest.mark.skip(reason="to_sympy() not implemented")
+    def test_against_sympy(self):
+        pass
+
 
 # class TestHarmonicOscillatorFromEquationUnits(EquationBase):
 #     Potential = from_equation("1/2*k*x**2", vars="x", pars="k",